add: graph type cast
implements a way to change graph type. e. g. from directed to undirected.

now it is possible to get the type of the graph for a user
