+++ library/src/main/kotlin/graph/implementation/UndirectedUnweightedGraph.kt
@@ -1,11 +1,5 @@
 package graph.implementation
 
-open class UndirectedUnweightedGraph<V, E> : UndirectedWeightedGraph<V, E>() {
-    override fun getEdgeWeight(e: E): Double {
-        return DEFAULT_EDGE_WEIGHT
-    }
+import graph.abstracts.AbstractGraph
 
-    override fun setEdgeWeight(e: E, w: Double) {
-        return
-    }
-}
+open class UndirectedUnweightedGraph<V, E> : AbstractGraph<V, E>(false, false)
