+++ library/src/main/kotlin/graph/implementation/UndirectedWeightedGraph.kt
@@ -1,38 +1,5 @@
 package graph.implementation
 
-open class UndirectedWeightedGraph<V, E> : DirectedWeightedGraph<V, E>() {
-    override fun containsEdge(tail: V, head: V): Boolean {
-        return super.containsEdge(tail, head) || super.containsEdge(head, tail)
-    }
+import graph.abstracts.AbstractGraph
 
-    override fun removeEdge(tail: V, head: V): Boolean {
-        val isPresent = super.removeEdge(tail, head) || super.removeEdge(head, tail)
-        return isPresent
-    }
-
-    override fun edgeSetOfVertecies(): Set<Pair<V, V>> {
-        val connections = mutableSetOf<Pair<V, V>>()
-        val vertecies = vertexSet()
-        for (v1 in vertecies) {
-            for (v2 in vertecies) {
-                if (graph.get(v1, v2) != null) {
-                    connections.add(Pair(v1, v2))
-                    connections.add(Pair(v2, v1))
-                }
-            }
-        }
-        return connections
-    }
-
-    override fun getEdge(tail: V, head: V): E? = graph.get(tail, head)?.data ?: graph.get(head, tail)?.data
-
-    override fun inDegreeOf(v: V): Int = degreeOf(v)
-
-    override fun outDegreeOf(v: V): Int = degreeOf(v)
-
-    override fun incomingEdgesOf(v: V): Set<E> = super.incomingEdgesOf(v) + super.outgoingEdgesOf(v)
-    override fun outgoingEdgesOf(v: V): Set<E> = super.incomingEdgesOf(v) + super.outgoingEdgesOf(v)
-    override fun incomingVerteciesOf(v: V): Set<V> = super.incomingVerteciesOf(v) + super.outgoingVerteciesOf(v)
-    override fun outgoingVerteciesOf(v: V): Set<V> = super.incomingVerteciesOf(v) + super.outgoingVerteciesOf(v)
-
-}
+open class UndirectedWeightedGraph<V, E> : AbstractGraph<V, E>(false, true)
